feat : 조회수 증분을 int 형식에서 long 형식으로 수정, 불필요한 type casting 줄이기 위해서

+++ src/main/java/com/board/post/entity/Post.java
@@ -71,7 +71,7 @@ public class Post extends BaseEntity {
         return this;
     }
 
-    public void incrementViews(int increment) {
+    public void incrementViews(long increment) {
         this.viewCount = this.viewCount + increment;
     }
 }
